Refactor import statements in GameCard and PlatformIconList
