+++ src/entities/Game.ts
@@ -1,0 +1,12 @@
+import { Platform } from "@/entities/Platform";
+
+export interface Game {
+  id: number;
+  name: string;
+  slug: string;
+  background_image: string;
+  parent_platforms: { platform: Platform }[];
+  metacritic: number;
+  rating_top: number;
+  description_raw: string;
+}
