+++ src/hooks/useGame.ts
@@ -1,6 +1,6 @@
 import APIClient from "@/services/api-client";
 import { keepPreviousData, useQuery } from "@tanstack/react-query";
-import { Game } from "./useGames";
+import { Game } from "../entities/Game";
 
 const apiClient = new APIClient<Game>("/games");
 const useGame = (slug: string) =>
