+++ src/hooks/useGames.ts
@@ -1,19 +1,9 @@
 import { keepPreviousData, useInfiniteQuery } from "@tanstack/react-query";
-import APIClient, { FetchResponse, Platform } from "@/services/api-client";
+import APIClient, { FetchResponse } from "@/services/api-client";
 import useGameQueryStore from "@/store";
+import { Game } from "../entities/Game";
 
 const apiClient = new APIClient<Game>("/games");
-
-export interface Game {
-  id: number;
-  name: string;
-  slug: string;
-  background_image: string;
-  parent_platforms: { platform: Platform }[];
-  metacritic: number;
-  rating_top: number;
-  description_raw: string;
-}
 
 const useGames = () => {
   const gameQuery = useGameQueryStore((s) => s.gameQuery);
