+++ src/hooks/useGenres.ts
@@ -1,13 +1,8 @@
 import APIClient from "@/services/api-client";
 import { useQuery } from "@tanstack/react-query";
+import { Genre } from "../entities/Genre";
 
 const apiClient = new APIClient<Genre>("/genres");
-
-export interface Genre {
-  id: number;
-  name: string;
-  image_background: string;
-}
 
 const useGenres = () => {
   return useQuery({
