+++ src/hooks/usePlatforms.tsx
@@ -1,5 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
-import APIClient, { Platform } from "@/services/api-client";
+import APIClient from "@/services/api-client";
+import { Platform } from "@/entities/Platform";
 
 const apiClient = new APIClient<Platform>("/platforms/lists/parents");
